+++ INTERX/tasks/cache_header_check.go
@@ -44,6 +44,7 @@ func CacheHeaderCheck(rpcAddr string, isLog bool) {
 					common.Mutex.Lock()
 					// check if file or path exists
 					if _, err := os.Stat(path); os.IsNotExist(err) {
+						common.Mutex.Unlock()
 						return nil
 					}
 					err := os.Remove(path)